+++ dragonaltar/src/main/java/com/dragonaltar/config/ServerMode.java
@@ -1,5 +1,5 @@
 package com.dragonaltar.config;
 
 public enum ServerMode {
-	BETA, PRODUCTION
+	DEVELOPMENT, BETA, PRODUCTION
 }